requirements.txt
face_recognition==1.3.0
feedparser==6.0.10
keyboard==0.13.5
nltk==3.7
numpy==1.23.4
opencv_python==4.6.0.66
Pillow==9.2.0
pytesseract==0.3.9
pyttsx3==2.90
pywin32==304
selenium==4.5.0
SpeechRecognition==3.8.1
tensorflow==2.9.1
webdriver_manager==3.8.3
